MIDRETEST 5-17-24.R
Linear regression is a statistical method used to model the relationship between a dependent variable and one or more independent variables. The goal is to find the best-fitting straight line (regression line) that predicts the dependent variable based on the values of the independent variables.

Polynomial regression is a type of regression analysis in which the relationship between the independent variable (or variables) and the dependent variable is modeled as an 
𝑛
n-th degree polynomial. Unlike simple linear regression, which fits a straight line to the data, polynomial regression can fit a curve to the data, making it useful for capturing non-linear relationships.

Logistic regression is a type of regression analysis used for predicting the outcome of a dependent variable that is categorical. It is commonly used for binary classification problems, where the dependent variable can take on one of two possible values, often coded as 0 or 1.